(feat) Refine file structure, including structure for modular tests from browser and/or terminal.
