+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "nodemon server/server.js",
     "postinstall": "bower install",
-    "test": "./node_modules/.bin/mocha --bail specs/test.js;"
+    "test": "./node_modules/.bin/mocha --bail specs/npmTests.js;"
   },
   "repository": {
     "type": "git",
