+++ specs/SpecRunner.html
@@ -10,13 +10,13 @@
   <script src="../node_modules/chai/chai.js"></script>
   <script src="../node_modules/mocha/mocha.js"></script>
 
-  <!-- Import Client Code Here -->
-  <script src="../js/functions.js"></script>
+  <!-- Import Client or Server Code Here -->
+  <script src="../client/app.js"></script>
 
   <script>mocha.setup('bdd')</script>
 
-  <!-- Import Client Tests Here -->
-  <script src="test.js"></script>
+  <!-- Import Tests Here -->
+  <script src="exampleTest.js"></script>
 
   <script>
     mocha.run();
