+++ specs/npmTests.js
@@ -1,0 +1,3 @@
+// Write your tests within a function set to module.exports
+// and run it below.
+require('../specs/exampleTest')();
